refactor: extract LaunchAtLoginStatus.swift from LaunchAtLoginService.swift
Splits the `LaunchAtLoginStatus` enum (~53 lines with title / detail /
severity accessors) out of LaunchAtLoginService.swift.

Byte-identical move of lines 4-56. New file needs only Foundation
(no ServiceManagement); LaunchAtLoginService.swift keeps both imports
for the SMAppService-backed implementations.

LaunchAtLoginService.swift: 112 → 58 lines. Tests: 667.
